CartoDBGroupLayer knows how to fetch attributes
